Simplify Vector
