+++ installation_instructions
@@ -4,7 +4,7 @@ module list
 ouput should have the following modules
 
   1) gcc/8.2.0      3) openmpi/4.0.6   5) cuda/11.4.0
-  2) mkl/2021.3.0   4) cmake/3.17.3    6) python3.6-anaconda/5.2.0 
+  2) mkl/2021.3.0   4) cmake/3.21.3    6) python3.6-anaconda/5.2.0 
 
 install sphinx, sphinx-rtd-theme and breathe using pip (optional)
