+++ src/linearAlgebra/CMakeLists.txt
@@ -9,12 +9,11 @@ if (NOT TARGET dft-efe-linalg)
             BlasLapack.cpp
             BlasLapackKernels.cpp
             BlasLapackKernels.cu
-            DistributedVector.cpp
             Matrix.cpp
             QueueManager.cpp
             LinAlgOpContext.cpp
-            SerialVector.cpp
             VectorAttributes.cpp
+	    Vector.cpp
             AbstractMatrix.cpp
             GeneralMatrix.cpp
             HermitianMatrix.cpp
