--- src/linearAlgebra/DistributedVector.cpp
@@ -1,1 +1,0 @@
-#include "DistributedVector.h"
