--- src/linearAlgebra/SerialVector.cpp
@@ -1,1 +1,0 @@
-#include "SerialVector.h"
