+++ src/utils/MPIWrapper.t.cpp
@@ -361,6 +361,7 @@ namespace dftefe
         return ::MPI_Issend(buf, count, datatype, dest, tag, comm, request);
       }
 
+
 #else // DFTEFE_WITH_MPI
 
       template <MemorySpace memorySpace>
